# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Unifiles API
  description: Canonical contract for the self-hosted Unifiles document processing API. Server bindings
    and official SDK cores are generated from this document.
  version: 1.0.0
paths:
...
  /v1/files:
    post:
      tags:
      - Files
      summary: Upload File
      operationId: uploadFile
      security:
      - BearerAuth: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
          title: Idempotency-Key
          nullable: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadFileRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      tags:
      - Files
      summary: List Files
      operationId: listFiles
      security:
      - BearerAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: tags
        in: query
        required: false
        schema:
          type: string
          title: Tags
          nullable: true
      - name: content_type
        in: query
        required: false
        schema:
          type: string
          title: Content Type
          nullable: true
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          default: created_at
          title: Sort By
      - name: order
        in: query
        required: false
        schema:
          type: string
          default: desc
          title: Order
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /v1/knowledge-bases/{kb_id}/documents:
    post:
      tags:
      - Documents
      summary: Create Document
      operationId: createDocument
      security:
      - BearerAuth: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: string
          title: Kb Id
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
          title: Idempotency-Key
          nullable: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreate'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      tags:
      - Documents
      summary: List Documents
      operationId: listDocuments
      security:
      - BearerAuth: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: string
          title: Kb Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/knowledge-bases/{kb_id}/documents/{document_id}:
    get:
      tags:
      - Documents
      summary: Get Document
      operationId: getDocument
      security:
      - BearerAuth: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: string
          title: Kb Id
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags:
      - Documents
      summary: Delete Document
      operationId: deleteDocument
      security:
      - BearerAuth: []
      parameters:
      - name: kb_id
        in: path
        required: true
        schema:
          type: string
          title: Kb Id
      - name: document_id
        in: path
        required: true
        schema:
          type: string
          title: Document Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Permission or quota denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Resource state conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    DeletionResult:
      properties:
        id:
          type: string
          title: Id
        deleted:
          type: boolean
          title: Deleted
          default: true
      additionalProperties: false
      type: object
      required:
      - id
      title: DeletionResult
    DocumentCreate:
      properties:
        file_id:
          type: string
          title: File Id
        title:
          type: string
          title: Title
          nullable: true
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      additionalProperties: false
      type: object
      required:
      - file_id
      title: DocumentCreate
    DocumentResource:
      properties:
        id:
          type: string
          title: Id
        kb_id:
          type: string
          title: Kb Id
        file_id:
          type: string
          title: File Id
        title:
          type: string
          title: Title
          nullable: true
        status:
          type: string
          enum:
          - pending
          - indexing
          - indexed
          - failed
          - cancelled
          title: Status
        chunk_count:
          type: integer
          title: Chunk Count
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
          nullable: true
        error:
          additionalProperties: true
          type: object
          title: Error
          nullable: true
        created_at:
          type: string
          format: date-time
          title: Created At
        indexed_at:
          type: string
          format: date-time
          title: Indexed At
          nullable: true
      additionalProperties: false
      type: object
      required:
      - id
      - kb_id
      - file_id
      - status
      - chunk_count
      - created_at
      title: DocumentResource
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
        request_id:
          type: string
          title: Request Id
      additionalProperties: false
      type: object
      required:
      - code
      - message
      - request_id
      title: ErrorDetail
    ErrorEnvelope:
      properties:
        success:
          type: boolean
          title: Success
          default: false
        error:
          $ref: '#/components/schemas/ErrorDetail'
      additionalProperties: false
      type: object
      required:
      - error
      title: ErrorEnvelope
...
    FileResource:
      properties:
        id:
          type: string
          title: Id
        filename:
          type: string
          title: Filename
        content_type:
          type: string
          title: Content Type
        size:
          type: integer
          title: Size
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        tags:
          items:
            type: string
          type: array
          title: Tags
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
          nullable: true
      additionalProperties: false
      type: object
      required:
      - id
      - filename
      - content_type
      - size
      - created_at
      title: FileResource
...
    UploadFileRequest:
      properties:
        file:
          type: string
          format: binary
          title: File
        metadata:
          type: string
          title: Metadata
          default: '{}'
        tags:
          type: string
          title: Tags
          default: '[]'
      type: object
      required:
      - file
      title: UploadFileRequest
...
    DocumentList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/DocumentResource'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
        has_more:
          type: boolean
          title: Has More
      additionalProperties: false
      type: object
      required:
      - items
      - total
      - limit
      - offset
      - has_more
      title: DocumentList
...
    FileList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/FileResource'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
        has_more:
          type: boolean
          title: Has More
      additionalProperties: false
      type: object
      required:
      - items
      - total
      - limit
      - offset
      - has_more
      title: FileList
...
    DeletionResponse:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          $ref: '#/components/schemas/DeletionResult'
      additionalProperties: false
      type: object
      required:
      - data
      title: DeletionResponse
    DocumentResponse:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          $ref: '#/components/schemas/DocumentResource'
      additionalProperties: false
      type: object
      required:
      - data
      title: DocumentResponse
...
    FileResponse:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          $ref: '#/components/schemas/FileResource'
      additionalProperties: false
      type: object
      required:
      - data
      title: FileResponse
...
    DocumentListResponse:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          $ref: '#/components/schemas/DocumentList'
      additionalProperties: false
      type: object
      required:
      - data
      title: DocumentListResponse
...
    FileListResponse:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          $ref: '#/components/schemas/FileList'
      additionalProperties: false
      type: object
      required:
      - data
      title: FileListResponse
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
servers:
- url: https://api.unifiles.dev
  description: Hosted API
- url: http://localhost:8088
  description: Local self-hosted API
...